+++ FrontEnd/src/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from './App';
+import './styles/global.css'
+
+ReactDOM.render(
+  <App />,
+  document.getElementById('root'),
+);
