+++ FrontEnd/src/pages/Cardapio.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import TableCardapio from '../components/TableCardapio';
+
+const Cardapio = () => {
+  return (
+    <TableCardapio/>
+  )
+}
+
+export default Cardapio;
