+++ FrontEnd/src/styles/Footer.css
@@ -1,0 +1,118 @@
+.contact {
+  background-color: var(--bluePE);
+  width: 100%;
+  margin-top: 55px;
+  padding: 50px;
+  color: var(--white);
+  display: flex;
+  justify-content: space-between;
+}
+
+.left h6 {
+  font-size: 18px;
+  margin-left: 20px;
+}
+
+.left h3 {
+  color: var(--yellowPE);
+}
+
+.imgLeft {
+  margin-right: 0;
+}
+
+.middle {
+  margin-bottom: -40px;
+  display: flex;
+  align-items: flex-end;
+}
+
+.middle {
+    text-align: center;
+    margin-top: 200px;
+    margin-left: 30px;
+}
+
+.middle h4 {
+    font-size: 20px;
+}
+
+.middle h4 a {
+    text-decoration: none;
+    color: var(--yellowPE);
+    transition: 0.2s ease-in-out;
+}
+
+.middle a:hover {
+    color: var(--greenPE);
+    text-decoration: overline;
+}
+
+.right {
+    text-align: right;
+}
+
+.right h6 {
+    font-size: 18px;
+    margin-right: 20px;
+}
+
+.right a {
+    text-decoration: none;
+    color: var(--yellowPE);
+    transition: 0.2s ease-in-out;
+}
+
+.right a:hover {
+    color: var(--greenPE);
+    text-decoration: overline;
+}
+
+.right img, svg {
+    vertical-align: middle;
+    margin-right: 20px;
+}
+
+@media (max-width: 1300px) {
+    .contact {
+        height: 500px;
+        align-items: center;
+        display: flex;
+        flex-direction: column;
+    }
+
+    .left {
+        text-align: center;
+        margin: 0 auto;
+        padding: 10px;
+    }
+
+    .left h6 {
+        margin-left: -10px;
+        font-size: 16px;
+    }
+
+    .middle {
+      margin: 30px;
+      text-align: center;
+      align-items: flex-end;
+    }
+
+    .middle h4 {
+      font-size: 16px;
+    }
+
+    .right {
+      text-align: center;
+    }
+
+    .right h6 {
+      font-size: 16px;
+    }
+}
+
+@media (max-width: 1300px){
+  .middle h4 {
+    font-size: 12px;
+  }
+}
