--- src/pages/HorarioRefeicoes.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import TableRedeicoes from '../components/TableRefeicoes';
-
-const HorarioRefeicoes = () => {
-  return (
-    <TableRedeicoes/>
-  )
-};
-
-export default HorarioRefeicoes;
