--- src/styles/FormSelects.css
@@ -1,21 +1,0 @@
-.options{
-  width: 100%;
-  max-width: 1000px;
-  margin: 0 auto;
-  margin-bottom: 20px;
-  display: flex;
-  justify-content: space-around;
-}
-
-@media (min-width: 1610px){
-  .options {
-    max-width: 1500px;
-  }
-}
-
-@media (max-width: 1100px) {
-  .options{
-    width: 95%;
-    margin-top: -25px;
-  }
-}
