--- src/styles/global.css
@@ -1,66 +1,0 @@
-:root {
-  --white: #fff;
-  --gray-100: #edf2f7;
-  --gray-200: #e2e8f0;
-  --gray-400: #a0aec0;
-  --gray-500: #718096;
-  --gray-700: #2d3748;
-  --primary: #4763ff;
-  --primary-light: #7086ff;
-  --primary-dark: #1834cc;
-  --danger: #e53e3e;
-  --danger-light: #f9eaea;
-  --bluePE: #3155A4;
-  --redPE: #C34342;
-  --yellowPE: #FFB511;
-  --greenPE: #00AD4A;
-}
-
-::-webkit-scrollbar {
-  width: 10px;
-  background-color: var(--white);
-}
-
-::-webkit-scrollbar-thumb {
-  height: 10px;
-  background-color: var(--greenPE);
-}
-
-::-webkit-scrollbar-button {
-  background-color: var(--greenPE);
-  height: 10px;
-}
-
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  font-family: 16px;
-  font-family: 'Poppins', sans-serif;
-  list-style: none;
-}
-
-body {
-  background-color: var(--white);
-  color: var(--gray-700);
-}
-
-.TitleText {
-  color: var(--bluePE);
-  opacity: 1;
-  font-size: 28px;
-  text-align: center;
-}
-
-.line {
-  margin-bottom: -1px;
-  width: 50%;
-  border: 1px solid var(--bluePE);
-  opacity: 1;
-}
-
-.lineMidlle {
-  width: 35%;
-  border: 1px solid var(--bluePE);
-  opacity: 1;
-}
